requirements.txt
pandas
numpy
scipy
scikit-learn
flask
implicit
scikit-surprise
